backend/requirements/common.txt
django~=3.0.0
djangorestframework~=3.11.2
djangorestframework-jwt
django-cors-headers
django-grappelli
django-pydenticon
drf-yasg
pillow
